+++ CHANGELOG.md
@@ -36,3 +36,4 @@
 - Genes are now an actual Registry
 - Claws inflicts Bleeding for 5 seconds rather than 300
 - Fixed the duplicate message send when you get a Gene that you're missing the required Genes for
+- Players with Meaty can now shear themselves while sneaking